Move files according to the project structure
